Renaming files to follow consistent naming convention.
