+++ Copter/TestScripts/sitl_takeoff.py
